# 8257772: Vectorizing clear memory operation using AVX-512 masked operations