… src/hotspot/cpu/x86/x86_64.ad
… src/hotspot/share/opto/memnode.cpp
… test/micro/org/openjdk/bench/vm/compiler/ClearMemory.java