@jatin-bhateja 2020-12-15
> Passed my tier1-tier3 testing with UseAVX=3.

Thanks @vnkozlov  and @TobiHartmann